refactor(ui): rewrite claude-extension page to design system
Extract 1118-LOC monolith into claude-extension/ folder:
- index.tsx: state + handlers only (~150 LOC)
- claude-extension-left-panel.tsx: sidebar with editor + list
- claude-extension-content-panel.tsx: main area tabs orchestrator
- sections/binding-editor-section.tsx: form fields + save/delete
- sections/binding-list-section.tsx: selectable saved binding rows
- sections/targets-section.tsx: shared + IDE TargetStatusCards
- sections/resolved-binding-section.tsx: resolved binding + payload
- sections/diagnostics-section.tsx: warnings + notes cards
- sections/advanced-section.tsx: raw JSON + metadata
- lib/claude-extension-draft-utils.ts: BindingDraft type + converters
- lib/claude-extension-ui-atoms.tsx: StatusBadge, DetailRow, CodeBlockCard, ErrorBanner

Zero behavior changes. All catalogue items preserved. Build + typecheck + lint clean.
